% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettize.R
\name{palettize}
\alias{palettize}
\title{Create a color palette}
\usage{
palettize(path)
}
\arguments{
\item{path}{A path to a supported image file.}
}
\value{
A character vector of hexadecimal colors.
}
\description{
\code{palettize()} creates a color palette from an image file.
}
