% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettizer-package.R
\docType{package}
\name{palettizer-package}
\alias{palettizer}
\alias{palettizer-package}
\title{palettizer: R Package for Color Palette Generation Using the 'palettize' C++ Library}
\description{
R package for color palette generation using the 'palettize' C++ library.
}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github@gmail.com}

}
\keyword{internal}
