% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_check.R
\name{plt_check}
\alias{plt_check}
\title{Check if a file is an image from a supported format}
\usage{
plt_check(path)
}
\arguments{
\item{path}{A path to a file to be checked.}
}
\value{
\code{plt_check()} returns \code{TRUE} if the file is an image from a supported format,
and returns \code{FALSE} for non-image files and images from unsupported formats.
}
\description{
\code{plt_check()} is a wrapper around the C++ function \code{stbi_info} from
\code{stb_image.h}. It checks whether a file is an image from a supported format.
This is useful for checking whether \code{plt_tize()} can create a color palette
from a file without having to decode the entire file.
}
\details{
Future versions of this function will expose how to query the width, height,
and component count of an image.
}
