% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_tize.R
\name{plt_tize}
\alias{plt_tize}
\title{Create a color palette}
\usage{
plt_tize(path)
}
\arguments{
\item{path}{A path to a supported image file.}
}
\value{
A character vector of hexadecimal colors.
}
\description{
\code{plt_tize()} creates a color palette from a supported image file.
}
